include: drivers: pwm: extend the pwm flags to be soc specific
Add 8 bits to the pwm_flags_t so that upper byte is reserved
for non-standard but soc specific dts flags and keeping the low
Byte for standard flags.
Some of SoC like STM32 mcus can then define their own flags
in their dt-bindings/pwm.

+++ include/drivers/pwm.h
@@ -64,8 +64,12 @@ extern "C" {
 
 /**
  * @brief Provides a type to hold PWM configuration flags.
+ *
+ * The lower 8 bits are used for standard flags.
+ * The upper 8 bits are reserved for SoC specific flags.
  */
-typedef uint8_t pwm_flags_t;
+
+typedef uint16_t pwm_flags_t;
 
 /**
  * @typedef pwm_pin_set_t
